Fixed misplaced left middle pseudokeys
This was actually an issue that `mypy` couldn't protect against, because
the iteration of strings returns... strings. What happened was that to
check if a pseudokey was part of the left middle keys, it checked if any
of the elements of it are physical middle keys, which in this case was
erroneously done by iterating over the string and checking the results,
instead of the correct iteration which is over the replacement for that
pseudokey - a list of strings.

Closes #13